Move ConversionError enum into the error house
